requirements.txt
pyautogui
google_play_scraper
fire
spotipy
python-dotenv
twilio
inputimeout